+++ shared/utils.h
@@ -4,7 +4,7 @@
 void write_weights_file(char *filename, float *weights, int num_weights);
 
 // This function is modified for CNN weights
-bool read_weights_file(const char *filename, float *weights, size_t weight_size);
+bool read_weights_file(const char *filename, float *weights, int weight_size);
 
 unsigned int convert_endian_4bytes(unsigned int input);
 int parse_MNIST_images(const char* file, unsigned char** X);
